various fixes of all kinds, Hexen still not working on gl-friendly segs

+++ src/hexen/r_local.h
@@ -219,8 +219,8 @@ typedef struct polyblock_s
 typedef struct subsector_s
 {
     sector_t *sector;
-    short numlines;
-    short firstline;
+    int numlines;  // [crispy] extended nodes
+    int firstline;  // [crispy] extended nodes
     polyobj_t *poly;
 } subsector_t;
 
@@ -228,7 +228,8 @@ typedef struct
 {
     fixed_t x, y, dx, dy;       // partition line
     fixed_t bbox[2][4];         // bounding box for each child
-    unsigned short children[2]; // if NF_SUBSECTOR its a subsector
+    // if NF_SUBSECTOR its a subsector
+    int children[2]; // [crispy] extended nodes
 } node_t;
